chore: move PodStatus to API package
