+++ api/v1/base_types.go
@@ -20,6 +20,20 @@ import (
 	corev1 "k8s.io/api/core/v1"
 )
 
+// PodStatus represent the possible status of pods
+type PodStatus string
+
+const (
+	// PodHealthy means that a Pod is active and ready
+	PodHealthy = "healthy"
+
+	// PodReplicating means that a Pod is still not ready but still active
+	PodReplicating = "replicating"
+
+	// PodFailed means that a Pod will not be scheduled again (deleted or evicted)
+	PodFailed = "failed"
+)
+
 // SecretKeySelectorToCore transforms a SecretKeySelector structure to the
 // analogue one in the corev1 namespace
 func SecretKeySelectorToCore(selector *SecretKeySelector) *corev1.SecretKeySelector {
