+++ pkg/reconciler/persistentvolumeclaim/status.go
@@ -139,7 +139,7 @@ func EnrichStatus(
 
 	filteredPods := utils.FilterActivePods(runningInstances)
 	cluster.Status.ReadyInstances = utils.CountReadyPods(filteredPods)
-	cluster.Status.InstancesStatus = utils.ListStatusPods(runningInstances)
+	cluster.Status.InstancesStatus = apiv1.ListStatusPods(runningInstances)
 
 	cluster.Status.PVCCount = int32(len(managedPVCs)) //nolint:gosec
 	cluster.Status.InitializingPVC = result.getSorted(initializing)
