+++ pkg/utils/pod_conditions_test.go
@@ -162,106 +162,4 @@ var _ = Describe("Pod conditions test suite", func() {
 		}
 		Expect(IsPodUnschedulable(pod)).To(BeFalse())
 	})
-
-	Describe("Properly builds ListStatusPods", func() {
-		healthyPod := corev1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: "healthyPod",
-			},
-			Status: corev1.PodStatus{
-				Conditions: []corev1.PodCondition{
-					{
-						Type:   corev1.ContainersReady,
-						Status: corev1.ConditionTrue,
-					},
-				},
-			},
-		}
-		activePod := corev1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: "activePod",
-			},
-			Status: corev1.PodStatus{
-				Phase: corev1.PodRunning,
-				Conditions: []corev1.PodCondition{
-					{
-						Type:   corev1.ContainersReady,
-						Status: corev1.ConditionFalse,
-					},
-				},
-			},
-		}
-		failedPod := corev1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: "failedPod",
-			},
-			Status: corev1.PodStatus{
-				Phase: corev1.PodFailed,
-				Conditions: []corev1.PodCondition{
-					{
-						Type:   corev1.ContainersReady,
-						Status: corev1.ConditionFalse,
-					},
-				},
-			},
-		}
-
-		now := metav1.Now()
-		terminatingPod := corev1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:              "terminatingPod",
-				DeletionTimestamp: &now,
-			},
-			Status: corev1.PodStatus{
-				Phase: corev1.PodRunning,
-				Conditions: []corev1.PodCondition{
-					{
-						Type:   corev1.ContainersReady,
-						Status: corev1.ConditionTrue,
-					},
-				},
-			},
-		}
-
-		It("Detects healthy pods", func() {
-			podList := []corev1.Pod{healthyPod, healthyPod}
-			expectedStatus := map[PodStatus][]string{
-				PodHealthy: {"healthyPod", "healthyPod"},
-			}
-			podStatus := ListStatusPods(podList)
-			Expect(podStatus).To(BeEquivalentTo(expectedStatus))
-		})
-
-		It("Detects active pods", func() {
-			podList := []corev1.Pod{healthyPod, activePod}
-			expectedStatus := map[PodStatus][]string{
-				PodHealthy:     {"healthyPod"},
-				PodReplicating: {"activePod"},
-			}
-			podStatus := ListStatusPods(podList)
-			Expect(podStatus).To(BeEquivalentTo(expectedStatus))
-		})
-
-		It("Detects failed pods", func() {
-			podList := []corev1.Pod{healthyPod, activePod, failedPod}
-			expectedStatus := map[PodStatus][]string{
-				PodHealthy:     {"healthyPod"},
-				PodReplicating: {"activePod"},
-				PodFailed:      {"failedPod"},
-			}
-			podStatus := ListStatusPods(podList)
-			Expect(podStatus).To(BeEquivalentTo(expectedStatus))
-		})
-
-		It("Excludes terminating pods", func() {
-			podList := []corev1.Pod{healthyPod, activePod, failedPod, terminatingPod}
-			expectedStatus := map[PodStatus][]string{
-				PodHealthy:     {"healthyPod"},
-				PodReplicating: {"activePod"},
-				PodFailed:      {"failedPod"},
-			}
-			podStatus := ListStatusPods(podList)
-			Expect(podStatus).To(BeEquivalentTo(expectedStatus))
-		})
-	})
 })
